+++ iridauploader/core/cli_entry.py
@@ -9,7 +9,7 @@ from iridauploader.model import DirectoryStatus
 
 from . import api_handler, parsing_handler, logger, exit_return
 
-VERSION_NUMBER = "0.4.2"
+VERSION_NUMBER = "0.4.3"
 
 
 def upload_run_single_entry(directory, force_upload=False, upload_assemblies=False):